requirements.txt
Django
Pillow
python-dateutil
django-filter